multiple agents and dynamic workflows
before we only had one type of agent which output a very narrowly
defined manual description. Now we have given the LLM a lot more freedom
to interpet the user and come up with a solution.

evaluator still requires a lot of work, but for now, the objective is
not clearly defined so shall leave it be.